fixed plotnresponses and updated its tests [skip ci] todo: for side-by-side and all components plot, fix the y-axes labels
